+++ packages/theme/src/theme/colors/index.ts
@@ -1,0 +1,2 @@
+export * from './proposed'
+export * from './types'