+++ packages/theme/src/theme/index.ts
@@ -1,5 +1,5 @@
 export * from './border'
-export * from './colors'
 export * from './theme'
 export * from './types'
 export * from './typography'
+export * from './colors'